BF+TST: Fix sorting, change 'walk' method to 'matches'.
The semantics are quite different from os.walk, so move away from
that name.

Sorting is done strictly by name in DFS order, with no
consideration for if the path is a directory or not. This allows
a list of plain paths (without stat info) to be put in the same
sorted order.